requirements.txt
python-dotenv
google-adk
litellm
requests
google-cloud-aiplatform
google-auth
google-api-python-client
uvicorn
google-cloud-aiplatform[agent_engines]
pydantic
Flask